+++ R/utils-pipe.R
@@ -1,0 +1,10 @@
+#' Pipe operator
+#'
+#' Reexported pipe originally from magrittr's pipe operator.
+#'
+#' @name %>%
+#' @rdname pipe
+#' @param lhs A value or the magrittr placeholder.
+#' @param rhs A function call using the magrittr semantics.
+#' @export
+NULL