+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(RcppMeCab)
+library(purrr)
 
 test_check("RcppMeCab")